fix: well organized contacts component structure
