--- apps/gauzy/src/app/pages/contacts/contact.module.ts
@@ -1,96 +1,0 @@
-import { NgModule } from '@angular/core';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import {
-	NbButtonModule,
-	NbCardModule,
-	NbDialogModule,
-	NbIconModule,
-	NbInputModule,
-	NbTooltipModule,
-	NbSelectModule,
-	NbToggleModule,
-	NbSpinnerModule,
-	NbStepperModule
-} from '@nebular/theme';
-import { NgSelectModule } from '@ng-select/ng-select';
-import { ThemeModule } from '../../@theme/theme.module';
-import { SharedModule } from '../../@shared/shared.module';
-import { FileUploaderModule } from '../../@shared/file-uploader-input/file-uploader-input.module';
-import { EntityWithMembersModule } from '../../@shared/entity-with-members-card/entity-with-members-card.module';
-import { TagsColorInputModule } from '../../@shared/tags/tags-color-input/tags-color-input.module';
-import { EmployeeMultiSelectModule } from '../../@shared/employee/employee-multi-select/employee-multi-select.module';
-import { ContactComponent } from './contact.component';
-import { ContactMutationComponent } from './contact-mutation/contact-mutation.component';
-import { ContactRoutingModule } from './contact-routing.module';
-import { OrganizationContactService } from '../../@core/services/organization-contact.service';
-import { OrganizationProjectsService } from '../../@core/services/organization-projects.service';
-import { InviteService } from '../../@core/services/invite.service';
-import { InviteContactComponent } from './invite-contact/invite-contact.component';
-import { CardGridModule } from '../../@shared/card-grid/card-grid.module';
-import { Ng2SmartTableModule } from 'ng2-smart-table';
-import { ClientsComponent } from './clients/clients.component';
-import { LeadsComponent } from './leads/leads.component';
-import { CustomersComponent } from './customers/customers.component';
-import { ImageUploaderModule } from '../../@shared/image-uploader/image-uploader.module';
-import { ContactActionComponent } from './table-components/contact-action/contact-action.component';
-import { LocationFormModule } from '../../@shared/forms/location';
-import { LeafletMapModule } from '../../@shared/forms/maps/leaflet/leaflet.module';
-import { TranslateModule } from '../../@shared/translate/translate.module';
-import { HeaderTitleModule } from '../../@shared/components/header-title/header-title.module';
-import { GauzyButtonActionModule } from '../../@shared/gauzy-button-action/gauzy-button-action.module';
-import { PaginationModule } from '../../@shared/pagination/pagination.module';
-
-const COMPONENTS = [
-	ContactComponent,
-	InviteContactComponent,
-	ContactMutationComponent,
-	ContactMutationComponent,
-	ClientsComponent,
-	CustomersComponent,
-	LeadsComponent,
-	ContactActionComponent
-];
-
-@NgModule({
-	imports: [
-		SharedModule,
-		ThemeModule,
-		ContactRoutingModule,
-		NbCardModule,
-		FormsModule,
-		ReactiveFormsModule,
-		NbButtonModule,
-		NbInputModule,
-		NbSelectModule,
-		NbIconModule,
-		NbDialogModule.forChild(),
-		NbTooltipModule,
-		NbStepperModule,
-		FileUploaderModule,
-		CardGridModule,
-		Ng2SmartTableModule,
-		NgSelectModule,
-		EntityWithMembersModule,
-		TagsColorInputModule,
-		Ng2SmartTableModule,
-		ImageUploaderModule,
-		EmployeeMultiSelectModule,
-		NbSpinnerModule,
-		TranslateModule,
-		NbToggleModule,
-		LocationFormModule,
-		LeafletMapModule,
-		HeaderTitleModule,
-		GauzyButtonActionModule,
-		PaginationModule,
-		CardGridModule
-	],
-
-	declarations: [...COMPONENTS],
-	providers: [
-		OrganizationContactService,
-		OrganizationProjectsService,
-		InviteService
-	]
-})
-export class ContactModule {}
