+++ apps/gauzy/src/app/pages/contacts/contacts.component.html
@@ -91,8 +91,7 @@
 			<ngx-contact-action
 				[rowData]="selectedOrganizationContact"
 				(updateResult)="onUpdateResult($event)"
-			>
-			</ngx-contact-action>
+			></ngx-contact-action>
 			<button
 				nbButton
 				class="action secondary"
