+++ apps/gauzy/src/app/pages/contacts/contacts.component.scss
